requirements.txt
requests==2.24.0
discord.py==1.6.0
beautifulsoup4==4.9.3
discord==1.0.1
python-dotenv==0.17.1
html5lib==1.1
